requirements.txt
selenium>=3.141.0
python-dotenv
requests
rich